refactor: create a new okhttp client for each target on jvm target
